feat(chapters): add update and delete flows
Adds updateChapter and useUpdateChapterMutation so chapters can be edited
alongside create, extracts the create/edit form into components/ChapterForm
and turns app/chapters/new.tsx into a routing shell that hydrates initial
values and selection when chapterId is present. The chapter detail screen
gets an ellipsis popover menu with Edit and Delete actions, mirroring the
moments and events patterns.

Made-with: Cursor
