+++ app/chapters/new.tsx
@@ -1,476 +1,185 @@
-import { DatePicker } from '@/components/DatePicker';
+import ChapterForm, {
+  type ChapterInitialSelection,
+} from '@/components/ChapterForm';
 import ModalTopBar from '@/components/ModalTopBar';
 import Divider from '@/components/ui/Divider';
-import { Dropdown } from '@/components/ui/Dropdown';
-import { Input } from '@/components/ui/Input';
 import { Text } from '@/components/ui/Text';
 import { useActiveGroup } from '@/hooks/useActiveGroup';
-import { useCreateChapterMutation } from '@/hooks/useChapters';
-import { useEventsQuery } from '@/hooks/useEvents';
-import { useMomentsQuery } from '@/hooks/useMoments';
-import {
-  type CreateChapterValues,
-  createChapterSchema,
-} from '@/lib/validation/chapter';
-import type { ChapterItemInput, ChapterItemKind } from '@/services/chapters';
+import { useChapterDetailQuery } from '@/hooks/useChapters';
+import { parseLocalDate } from '@/lib/date';
+import { type CreateChapterValues } from '@/lib/validation/chapter';
+import type { ChapterDetail } from '@/services/chapters';
 import { baseColors, sectionColors } from '@/theme/colors';
-import { radius } from '@/theme/radius';
 import { space } from '@/theme/space';
 import { text as textTheme } from '@/theme/type';
-import { useForm } from '@tanstack/react-form';
-import { Image } from 'expo-image';
-import { router } from 'expo-router';
-import { Check } from 'lucide-react-native';
-import { useMemo, useState } from 'react';
+import {
+  router,
+  useGlobalSearchParams,
+  useLocalSearchParams,
+} from 'expo-router';
 import {
   ActivityIndicator,
   KeyboardAvoidingView,
   Platform,
-  Pressable,
-  ScrollView,
   StyleSheet,
   View,
 } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
-const FALLBACK_COVER_IMAGE = require('../../assets/images/fallbackImage.png');
-
-function formatShortDate(value: string): string {
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return value;
-  }
-  return date.toLocaleDateString(undefined, {
-    month: 'short',
-    day: 'numeric',
-    year: 'numeric',
-  });
-}
-
-function makeItemKey(kind: ChapterItemKind, id: string) {
-  return `${kind}:${id}`;
-}
-
-type SelectableItem = {
-  kind: ChapterItemKind;
-  id: string;
-  title: string;
-  occurredOn: string;
-  coverImage?: string;
+const EMPTY_VALUES: CreateChapterValues = {
+  chapterType: '',
+  title: '',
+  description: '',
+  occurredAt: new Date(),
 };
 
-const chapterTypeOptions = [
-  { label: 'Vacation', value: 'vacation' },
-  { label: 'Trip', value: 'trip' },
-  { label: 'Project', value: 'project' },
-  { label: 'Season', value: 'season' },
-  { label: 'Milestone', value: 'milestone' },
-  { label: 'Celebration', value: 'celebration' },
-  { label: 'Miscellaneous', value: 'miscellaneous' },
-];
+function getParamValue(value?: string | string[]): string | undefined {
+  return Array.isArray(value) ? value[0] : value;
+}
+
+function getInitialValues(chapter: ChapterDetail): CreateChapterValues {
+  return {
+    chapterType: chapter.chapterType ?? '',
+    title: chapter.title,
+    description: chapter.description ?? '',
+    occurredAt: parseLocalDate(chapter.occurredOn),
+  };
+}
+
+function getInitialSelection(
+  chapter: ChapterDetail,
+): ChapterInitialSelection[] {
+  return chapter.entries.map((entry) => ({
+    kind: entry.kind,
+    id: entry.id,
+  }));
+}
+
+function getLoadError(options: {
+  activeGroupId: string | null | undefined;
+  error: unknown;
+  hasChapter: boolean;
+  isEdit: boolean;
+  isLoadingGroups: boolean;
+  isChapterPending: boolean;
+}): string | null {
+  const {
+    activeGroupId,
+    error,
+    hasChapter,
+    isEdit,
+    isLoadingGroups,
+    isChapterPending,
+  } = options;
+
+  if (!isEdit) {
+    return null;
+  }
+
+  if (!isLoadingGroups && !activeGroupId) {
+    return 'Choose a space before editing this chapter.';
+  }
+
+  if (error instanceof Error) {
+    return error.message;
+  }
+
+  if (error) {
+    return 'Could not load chapter.';
+  }
+
+  if (!isLoadingGroups && !isChapterPending && !hasChapter) {
+    return 'Chapter not found.';
+  }
+
+  return null;
+}
 
 export default function NewChapterScreen() {
-  const { activeGroup } = useActiveGroup();
-  const [selectedKeys, setSelectedKeys] = useState<Set<string>>(new Set());
-  const [selectionOrder, setSelectionOrder] = useState<string[]>([]);
-  const [hasTriedSubmit, setHasTriedSubmit] = useState(false);
-  const [isSaving, setIsSaving] = useState(false);
-  const [saveError, setSaveError] = useState<string | null>(null);
-  const [itemsError, setItemsError] = useState<string | null>(null);
-  const createChapterMutation = useCreateChapterMutation();
-  const momentsQuery = useMomentsQuery(activeGroup?.id);
-  const eventsQuery = useEventsQuery(activeGroup?.id);
-
-  const moments = momentsQuery.data ?? [];
-  const events = eventsQuery.data ?? [];
-  const isLoadingItems = momentsQuery.isPending || eventsQuery.isPending;
-  const itemsLoadError =
-    momentsQuery.error instanceof Error
-      ? momentsQuery.error.message
-      : eventsQuery.error instanceof Error
-        ? eventsQuery.error.message
-        : null;
-
-  const itemsByKey = useMemo(() => {
-    const map = new Map<string, SelectableItem>();
-    for (const moment of moments) {
-      const key = makeItemKey('moment', moment.id);
-      map.set(key, {
-        kind: 'moment',
-        id: moment.id,
-        title: moment.title,
-        occurredOn: moment.occurredOn,
-        coverImage: moment.coverImage,
-      });
-    }
-    for (const event of events) {
-      const key = makeItemKey('event', event.id);
-      map.set(key, {
-        kind: 'event',
-        id: event.id,
-        title: event.title,
-        occurredOn: event.occurredOn,
-        coverImage: event.coverImage,
-      });
-    }
-    return map;
-  }, [events, moments]);
-
-  const selectedCount = selectedKeys.size;
-
-  const form = useForm({
-    defaultValues: {
-      chapterType: '',
-      title: '',
-      description: '',
-      occurredAt: new Date(),
-    } satisfies CreateChapterValues,
-    onSubmit: async ({ value }) => {
-      const parsed = createChapterSchema.safeParse(value);
-      if (!parsed.success) {
-        throw new Error(parsed.error.issues[0]?.message ?? 'Invalid input.');
-      }
-
-      if (!activeGroup?.id) {
-        throw new Error('Choose a space before saving this chapter.');
-      }
-
-      const items: ChapterItemInput[] = selectionOrder
-        .map((key) => itemsByKey.get(key))
-        .filter((item): item is SelectableItem => Boolean(item))
-        .map((item) => ({ kind: item.kind, id: item.id }));
-
-      if (items.length === 0) {
-        throw new Error(
-          'Pick at least one moment or event to bundle into this chapter.',
-        );
-      }
-
-      try {
-        await createChapterMutation.mutateAsync({
-          ...parsed.data,
-          groupId: activeGroup.id,
-          items,
-        });
-      } catch (error) {
-        if (error instanceof Error) {
-          throw error;
-        }
-        throw new Error('Failed to save chapter. Please try again.');
-      }
-
-      router.back();
-    },
+  const { activeGroup, isLoading: isLoadingGroups } = useActiveGroup();
+  const localParams = useLocalSearchParams<{
+    chapterId?: string | string[];
+    id?: string | string[];
+  }>();
+  const globalParams = useGlobalSearchParams<{
+    chapterId?: string | string[];
+    id?: string | string[];
+  }>();
+  const chapterId = getParamValue(
+    localParams.chapterId ??
+      globalParams.chapterId ??
+      localParams.id ??
+      globalParams.id,
+  );
+  const isEdit = Boolean(chapterId);
+  const chapterQuery = useChapterDetailQuery(chapterId, activeGroup?.id);
+  const chapter = chapterQuery.data;
+  const isLoadingChapter =
+    isEdit && (isLoadingGroups || chapterQuery.isPending);
+  const loadError = getLoadError({
+    activeGroupId: activeGroup?.id,
+    error: chapterQuery.error,
+    hasChapter: Boolean(chapter),
+    isEdit,
+    isLoadingGroups,
+    isChapterPending: chapterQuery.isPending,
   });
 
-  const validation = useMemo(
-    () => createChapterSchema.safeParse(form.state.values),
-    [form.state.values],
-  );
-  const fieldErrors = validation.success
-    ? {}
-    : validation.error.flatten().fieldErrors;
-  const submitError = saveError ?? form.state.errorMap.onSubmit;
-
-  function toggleItem(item: SelectableItem) {
-    const key = makeItemKey(item.kind, item.id);
-    setSelectedKeys((current) => {
-      const next = new Set(current);
-      if (next.has(key)) {
-        next.delete(key);
-      } else {
-        next.add(key);
-      }
-      return next;
-    });
-    setSelectionOrder((current) => {
-      if (current.includes(key)) {
-        return current.filter((existing) => existing !== key);
-      }
-      return [...current, key];
-    });
-    setItemsError(null);
+  if (isLoadingChapter) {
+    return (
+      <SafeAreaView style={styles.screen} edges={['top']}>
+        <KeyboardAvoidingView
+          style={styles.screen}
+          behavior={Platform.OS === 'ios' ? 'padding' : undefined}
+        >
+          <ModalTopBar
+            color={sectionColors.chapters}
+            onClose={() => router.back()}
+            title="Edit Chapter"
+          />
+          <Divider color={sectionColors.chapters} />
+          <View style={styles.loadingState}>
+            <ActivityIndicator color={sectionColors.chapters} />
+            <Text style={styles.submittingText}>Loading chapter...</Text>
+          </View>
+        </KeyboardAvoidingView>
+      </SafeAreaView>
+    );
   }
 
-  const handleSave = async () => {
-    if (isSaving) {
-      return;
-    }
-
-    setHasTriedSubmit(true);
-    setSaveError(null);
-    setItemsError(null);
-    const parsed = createChapterSchema.safeParse(form.state.values);
-    if (!parsed.success) {
-      return;
-    }
-
-    if (!activeGroup?.id) {
-      setSaveError('Choose a space before saving this chapter.');
-      return;
-    }
-
-    if (selectedKeys.size === 0) {
-      setItemsError(
-        'Pick at least one moment or event to bundle into this chapter.',
-      );
-      return;
-    }
-
-    setIsSaving(true);
-
-    try {
-      await form.handleSubmit();
-    } catch (error) {
-      if (error instanceof Error) {
-        setSaveError(error.message);
-      } else {
-        setSaveError('Failed to save chapter. Please try again.');
-      }
-      setIsSaving(false);
-    }
-  };
-
-  const hasMoments = moments.length > 0;
-  const hasEvents = events.length > 0;
-  const hasAnyItems = hasMoments || hasEvents;
-
-  return (
-    <SafeAreaView style={styles.screen} edges={['top']}>
-      <KeyboardAvoidingView
-        style={styles.screen}
-        behavior={Platform.OS === 'ios' ? 'padding' : undefined}
-      >
-        <ModalTopBar
-          color={sectionColors.chapters}
-          onClose={() => router.back()}
-          onSave={handleSave}
-          title="New Chapter"
-        />
-        <Divider color={sectionColors.chapters} />
-
-        <ScrollView
-          automaticallyAdjustKeyboardInsets
-          contentContainerStyle={styles.content}
-          contentInsetAdjustmentBehavior="automatic"
-          keyboardDismissMode="interactive"
-          keyboardShouldPersistTaps="handled"
-          showsVerticalScrollIndicator={false}
+  if (loadError) {
+    return (
+      <SafeAreaView style={styles.screen} edges={['top']}>
+        <KeyboardAvoidingView
+          style={styles.screen}
+          behavior={Platform.OS === 'ios' ? 'padding' : undefined}
         >
-          <View style={styles.form}>
-            <form.Field name="chapterType">
-              {(field) => (
-                <Dropdown
-                  color={sectionColors.chapters}
-                  onChange={field.handleChange}
-                  options={chapterTypeOptions}
-                  placeholder="Chapter type"
-                  value={field.state.value || undefined}
-                />
-              )}
-            </form.Field>
-            {hasTriedSubmit && fieldErrors.chapterType?.[0] ? (
-              <Text style={styles.errorText}>{fieldErrors.chapterType[0]}</Text>
-            ) : null}
-
-            <form.Field name="title">
-              {(field) => (
-                <Input
-                  label="Title"
-                  onBlur={field.handleBlur}
-                  onChangeText={field.handleChange}
-                  placeholder="Name this chapter of your story..."
-                  value={field.state.value}
-                />
-              )}
-            </form.Field>
-            {hasTriedSubmit && fieldErrors.title?.[0] ? (
-              <Text style={styles.errorText}>{fieldErrors.title[0]}</Text>
-            ) : null}
-
-            <form.Field name="occurredAt">
-              {(field) => (
-                <DatePicker
-                  color={sectionColors.chapters}
-                  label="Date"
-                  onChange={field.handleChange}
-                  required
-                  value={field.state.value}
-                />
-              )}
-            </form.Field>
-            {hasTriedSubmit && fieldErrors.occurredAt?.[0] ? (
-              <Text style={styles.errorText}>{fieldErrors.occurredAt[0]}</Text>
-            ) : null}
-
-            <form.Field name="description">
-              {(field) => (
-                <Input
-                  label="Description"
-                  minRows={4}
-                  onBlur={field.handleBlur}
-                  onChangeText={field.handleChange}
-                  placeholder="What ties these memories together?"
-                  value={field.state.value}
-                  variant="textarea"
-                />
-              )}
-            </form.Field>
-            {hasTriedSubmit && fieldErrors.description?.[0] ? (
-              <Text style={styles.errorText}>{fieldErrors.description[0]}</Text>
-            ) : null}
-
-            <View style={styles.selectionSection}>
-              <View style={styles.selectionHeader}>
-                <Text style={styles.selectionLabel}>Memories</Text>
-                {selectedCount > 0 ? (
-                  <Text style={styles.selectionCount}>
-                    {selectedCount} selected
-                  </Text>
-                ) : null}
-              </View>
-              <Text style={styles.selectionHint}>
-                Pull existing moments and events into this chapter.
-              </Text>
-
-              {isLoadingItems ? (
-                <ActivityIndicator
-                  color={sectionColors.chapters}
-                  style={styles.sectionLoader}
-                />
-              ) : null}
-
-              {itemsLoadError ? (
-                <Text style={styles.errorText}>{itemsLoadError}</Text>
-              ) : null}
-
-              {!isLoadingItems && !itemsLoadError && !hasAnyItems ? (
-                <Text style={styles.selectionEmpty}>
-                  No moments or events to bundle yet. Create a few first, then
-                  come back.
-                </Text>
-              ) : null}
-
-              {!isLoadingItems && hasMoments ? (
-                <View style={styles.group}>
-                  <Text style={styles.groupLabel}>Moments</Text>
-                  <View style={styles.groupList}>
-                    {moments.map((moment) => {
-                      const item: SelectableItem = {
-                        kind: 'moment',
-                        id: moment.id,
-                        title: moment.title,
-                        occurredOn: moment.occurredOn,
-                        coverImage: moment.coverImage,
-                      };
-                      const key = makeItemKey('moment', moment.id);
-                      const selected = selectedKeys.has(key);
-                      return (
-                        <SelectableRow
-                          key={key}
-                          item={item}
-                          onToggle={toggleItem}
-                          selected={selected}
-                        />
-                      );
-                    })}
-                  </View>
-                </View>
-              ) : null}
-
-              {!isLoadingItems && hasEvents ? (
-                <View style={styles.group}>
-                  <Text style={styles.groupLabel}>Events</Text>
-                  <View style={styles.groupList}>
-                    {events.map((event) => {
-                      const item: SelectableItem = {
-                        kind: 'event',
-                        id: event.id,
-                        title: event.title,
-                        occurredOn: event.occurredOn,
-                        coverImage: event.coverImage,
-                      };
-                      const key = makeItemKey('event', event.id);
-                      const selected = selectedKeys.has(key);
-                      return (
-                        <SelectableRow
-                          key={key}
-                          item={item}
-                          onToggle={toggleItem}
-                          selected={selected}
-                        />
-                      );
-                    })}
-                  </View>
-                </View>
-              ) : null}
-
-              {hasTriedSubmit && itemsError ? (
-                <Text style={styles.errorText}>{itemsError}</Text>
-              ) : null}
-            </View>
-
-            {submitError ? (
-              <Text style={styles.errorText}>{submitError}</Text>
-            ) : null}
-
-            {isSaving ? (
-              <View style={styles.submittingRow}>
-                <ActivityIndicator color={sectionColors.chapters} />
-                <Text style={styles.submittingText}>Saving chapter...</Text>
-              </View>
-            ) : null}
+          <ModalTopBar
+            color={sectionColors.chapters}
+            onClose={() => router.back()}
+            title="Edit Chapter"
+          />
+          <Divider color={sectionColors.chapters} />
+          <View style={styles.loadingState}>
+            <Text style={styles.errorText}>{loadError}</Text>
           </View>
-        </ScrollView>
-      </KeyboardAvoidingView>
-    </SafeAreaView>
-  );
-}
+        </KeyboardAvoidingView>
+      </SafeAreaView>
+    );
+  }
 
-type SelectableRowProps = {
-  item: SelectableItem;
-  selected: boolean;
-  onToggle: (item: SelectableItem) => void;
-};
+  const initialValues = chapter ? getInitialValues(chapter) : EMPTY_VALUES;
+  const initialSelection = chapter ? getInitialSelection(chapter) : [];
 
-function SelectableRow({ item, onToggle, selected }: SelectableRowProps) {
   return (
-    <Pressable
-      accessibilityHint="Toggles this memory for the chapter"
-      accessibilityRole="button"
-      accessibilityState={{ selected }}
-      onPress={() => onToggle(item)}
-      style={({ pressed }) => [
-        styles.row,
-        selected ? styles.rowSelected : null,
-        pressed ? styles.rowPressed : null,
-      ]}
-    >
-      <View style={styles.thumbWrap}>
-        <Image
-          contentFit="cover"
-          source={
-            item.coverImage ? { uri: item.coverImage } : FALLBACK_COVER_IMAGE
-          }
-          style={styles.thumbImage}
-        />
-      </View>
-
-      <View style={styles.rowBody}>
-        <Text numberOfLines={1} style={styles.rowTitle}>
-          {item.title}
-        </Text>
-        <Text style={styles.rowMeta}>{formatShortDate(item.occurredOn)}</Text>
-      </View>
-
-      <View
-        style={[styles.checkbox, selected ? styles.checkboxSelected : null]}
-      >
-        {selected ? (
-          <Check color={baseColors.bg} size={14} strokeWidth={3} />
-        ) : null}
-      </View>
-    </Pressable>
+    <ChapterForm
+      key={`${chapterId ?? 'new'}:${activeGroup?.id ?? 'no-group'}`}
+      activeGroupId={activeGroup?.id ?? null}
+      chapterId={chapterId}
+      initialSelection={initialSelection}
+      initialValues={initialValues}
+      isEdit={isEdit}
+    />
   );
 }
 
@@ -479,15 +188,6 @@ const styles = StyleSheet.create({
     flex: 1,
     backgroundColor: baseColors.bg,
   },
-  content: {
-    paddingHorizontal: space.lg,
-    paddingTop: space.lg,
-    paddingBottom: space.xxl,
-    gap: space.xl,
-  },
-  form: {
-    gap: space.xl,
-  },
   errorText: {
     color: baseColors.textError,
     fontFamily: textTheme.family.medium,
@@ -495,121 +195,17 @@ const styles = StyleSheet.create({
     lineHeight: textTheme.lineHeight.xs,
     marginTop: -space.md,
   },
-  submittingRow: {
+  loadingState: {
     alignItems: 'center',
-    flexDirection: 'row',
+    flex: 1,
     gap: space.sm,
+    justifyContent: 'center',
+    paddingHorizontal: space.lg,
   },
   submittingText: {
     color: baseColors.textSoft,
     fontFamily: textTheme.family.medium,
     fontSize: textTheme.size.sm,
     lineHeight: textTheme.lineHeight.sm,
-  },
-  selectionSection: {
-    gap: space.md,
-  },
-  selectionHeader: {
-    alignItems: 'center',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  selectionLabel: {
-    color: baseColors.text,
-    fontFamily: textTheme.family.semiBold,
-    fontSize: textTheme.size.lg,
-    lineHeight: textTheme.lineHeight.md,
-  },
-  selectionCount: {
-    color: sectionColors.chapters,
-    fontFamily: textTheme.family.semiBold,
-    fontSize: textTheme.size.xs,
-    lineHeight: textTheme.lineHeight.xs,
-  },
-  selectionHint: {
-    color: baseColors.textSoft,
-    fontFamily: textTheme.family.regular,
-    fontSize: textTheme.size.xs,
-    lineHeight: textTheme.lineHeight.xs,
-  },
-  selectionEmpty: {
-    color: baseColors.textSoft,
-    fontFamily: textTheme.family.regular,
-    fontSize: textTheme.size.sm,
-    lineHeight: textTheme.lineHeight.sm,
-    paddingVertical: space.md,
-    textAlign: 'center',
-  },
-  sectionLoader: {
-    marginVertical: space.md,
-  },
-  group: {
-    gap: space.sm,
-  },
-  groupLabel: {
-    color: baseColors.textSoft,
-    fontFamily: textTheme.family.semiBold,
-    fontSize: textTheme.size.xs,
-    letterSpacing: 0.78,
-    lineHeight: textTheme.lineHeight.xs,
-    textTransform: 'uppercase',
-  },
-  groupList: {
-    gap: space.xs,
-  },
-  row: {
-    alignItems: 'center',
-    backgroundColor: baseColors.card,
-    borderColor: 'transparent',
-    borderRadius: radius.lg,
-    borderWidth: 1,
-    flexDirection: 'row',
-    gap: space.md,
-    padding: space.sm,
-  },
-  rowSelected: {
-    borderColor: sectionColors.chapters,
-  },
-  rowPressed: {
-    opacity: 0.9,
-  },
-  thumbWrap: {
-    borderRadius: radius.md,
-    height: 44,
-    overflow: 'hidden',
-    width: 44,
-  },
-  thumbImage: {
-    height: '100%',
-    width: '100%',
-  },
-  rowBody: {
-    flex: 1,
-    gap: 2,
-  },
-  rowTitle: {
-    color: baseColors.text,
-    fontFamily: textTheme.family.semiBold,
-    fontSize: textTheme.size.sm,
-    lineHeight: textTheme.lineHeight.sm,
-  },
-  rowMeta: {
-    color: baseColors.textSoft,
-    fontFamily: textTheme.family.regular,
-    fontSize: textTheme.size.xs,
-    lineHeight: textTheme.lineHeight.xs,
-  },
-  checkbox: {
-    alignItems: 'center',
-    borderColor: baseColors.textMuted,
-    borderRadius: radius.full,
-    borderWidth: 1.5,
-    height: 24,
-    justifyContent: 'center',
-    width: 24,
-  },
-  checkboxSelected: {
-    backgroundColor: sectionColors.chapters,
-    borderColor: sectionColors.chapters,
   },
 });
